alx_travel_app/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.1
django-admin==2.0.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.16.0
drf-yasg==1.21.10
excel-base==1.0.4
inflection==0.5.1
isoweek==1.3.3
kombu==5.5.3
mysqlclient==2.2.7
packaging==25.0
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
screen==1.0.1
six==1.17.0
sqlparse==0.5.3
TimeConvert==3.0.13
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
xlwt==1.3.0
